requirements.txt
asgiref==3.8.1
build==1.2.1
CacheControl==0.14.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.8
Django==5.0.6
django-environ==0.11.2
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.15.4
idna==3.7
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
more-itertools==10.3.0
msgpack==1.0.8
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.1.0
python-dotenv==1.0.1
rapidfuzz==3.9.3
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
sqlparse==0.5.0
tomlkit==0.12.5
trove-classifiers==2024.5.22
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
xattr==1.1.0
